+++ .dev-tools-dist/README.md
@@ -1,0 +1,3 @@
+**NOTE**: This folder contains the compiled output of the `dev-tools/` directory.
+You should be editing the `.ts` files there and then running `node build full` to
+force these `.js` files to be recreated.
