+++ package.json
@@ -26,7 +26,7 @@
     "pretest": "npm run lint && npm run build",
     "test": "mocha",
     "posttest": "npm run tsc",
-    "fulltest": "npm run pretest && mocha -g '$^' && npm run posttest"
+    "fulltest": "npm run lint && node build full && mocha -g '$^' && npm run posttest"
   },
   "husky": {
     "hooks": {
