--- test/dev-tools/harness.js
@@ -1,12 +1,0 @@
-'use strict';
-
-const assert = require('assert');
-
-const {MultiRunner} = require('./../../dev-tools/harness');
-
-describe('MultiRunner (slow)', async () => {
-	it('should run successfully', async () => {
-		const opts = {totalGames: 1, prng: [1, 2, 3, 4]};
-		assert.strictEqual(await (new MultiRunner(opts).run()), 0);
-	});
-});
